New: Configurable log file size limit
