+++ src/NzbDrone.Common/Options/LogOptions.cs
@@ -5,6 +5,7 @@ public class LogOptions
     public string Level { get; set; }
     public bool? FilterSentryEvents { get; set; }
     public int? Rotate { get; set; }
+    public int? SizeLimit { get; set; }
     public bool? Sql { get; set; }
     public string ConsoleLevel { get; set; }
     public string ConsoleFormat { get; set; }
